+++ sdk/openai.go
@@ -3,9 +3,11 @@ package sdk
 import (
 	"encoding/json"
 	"errors"
+	"fmt"
 	"io"
 	"net/http"
 
+	"github.com/LordPax/aicli/lang"
 	"github.com/LordPax/aicli/utils"
 )
 
@@ -132,4 +134,9 @@ func (o *OpenaiText) SendRequest(text string) (Message, error) {
 	}
 
 	return respMessage, nil
+}
+
+func (t *OpenaiText) AppendImageHistory(role, fileType string, file []byte) error {
+	l := lang.GetLocalize()
+	return fmt.Errorf(l.Get("not-implemented"), "AppendImageHistory")
 }