+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+* Parameter -f in text command accept image file (only for claude sdk)
+
 ## [0.4.0] - 2024-09-28
 
 ### Changed